Close #507 get image-url with Faraday
